Fix image example
Changed optimization level to s, which produces slightly larger
binaries, but fixes a miscompilation with the image example.

In summary using it seems that using opt-level 'z' causes a
miscompilation of the `inflate` crate, which is used to decode png
images in the `image` crate required for the `image` druid example.
The precise location that causes an index out of bounds error in the inflate
crate seems to be the return point at:
https://github.com/image-rs/inflate/blob/5a8c97984bcb5701f3d7111f28af85e3bfa0c927/src/lib.rs#L952

Additionally we had to load the image at compile time and embed it
directly into the wasm binary instead of loading it at run-time.